+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3]  - 2021-01-04
+
+### Fixed
+
+- Bug fix with FTP engine when FTP is not configured to use jail
+
 ## [2.0.2]  - 2020-10-24
 
 ### Fixed